+++ apps/web/src/components/screen-size-indicator.tsx
@@ -1,6 +1,6 @@
 import { Icons } from "@/components/icons"
 
-import { env } from "@soli/env/web/server"
+import { env } from "@solistack/env/web/server"
 
 export const ScreenSizeIndicator = () => {
   if (env.NODE_ENV === "production") return null