+++ config/tailwind/src/preset.ts
@@ -13,7 +13,7 @@ export { extract } from "fluid-tailwind"
 
 const colors = createPlugin()
 
-export const soliTailwindPreset: Config = {
+export const solistackTailwindPreset: Config = {
   content: { files: ["./src/**/*.{js,jsx,ts,tsx,mdx}"] },
   presets: [windyTypography],
   theme: {