+++ packages/db/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @soli/db
+# @solistack/db
 
 ## 0.1.0